# [Serverless] Make educated first guess when getting event type.