--- a/pkg/serverless/trigger/events.go
+++ b/pkg/serverless/trigger/events.go
@@ -17,8 +17,11 @@ import (
 type AWSEventType int
 
 const (
+	// Unknown describes an unknown event type
+	Unknown AWSEventType = iota
+
 	// APIGatewayEvent describes an event from classic AWS API Gateways
-	APIGatewayEvent AWSEventType = iota
+	APIGatewayEvent
 
 	// APIGatewayV2Event describes an event from AWS API Gateways
 	APIGatewayV2Event
@@ -70,97 +73,66 @@ const (
 
 	// LambdaFunctionURLEvent describes an event from an HTTP lambda function URL invocation
 	LambdaFunctionURLEvent
-
-	// Unknown describes an unknown event type
-	Unknown
 )
 
 // eventParseFunc defines the signature of AWS event parsing functions
 type eventParseFunc func(map[string]any) bool
 
+type checker struct {
+	check eventParseFunc
+	typ   AWSEventType
+}
+
+var (
+	unknownChecker   = new(checker)
+	lastEventChecker = unknownChecker
+	eventCheckers    = []*checker{
+		{isAPIGatewayEvent, APIGatewayEvent},
+		{isAPIGatewayV2Event, APIGatewayV2Event},
+		{isAPIGatewayWebsocketEvent, APIGatewayWebsocketEvent},
+		{isAPIGatewayLambdaAuthorizerTokenEvent, APIGatewayLambdaAuthorizerTokenEvent},
+		{isAPIGatewayLambdaAuthorizerRequestParametersEvent, APIGatewayLambdaAuthorizerRequestParametersEvent},
+		{isALBEvent, ALBEvent},
+		{isCloudFrontRequestEvent, CloudFrontRequestEvent},
+		{isCloudwatchEvent, CloudWatchEvent},
+		{isCloudwatchLogsEvent, CloudWatchLogsEvent},
+		{isDynamoDBStreamEvent, DynamoDBStreamEvent},
+		{isKinesisStreamEvent, KinesisStreamEvent},
+		{isS3Event, S3Event},
+		{isSNSEvent, SNSEvent},
+		{isSNSSQSEvent, SNSSQSEvent},
+		{isSQSEvent, SQSEvent},
+		{isAppSyncResolverEvent, AppSyncResolverEvent},
+		{isEventBridgeEvent, EventBridgeEvent},
+		{isLambdaFunctionURLEvent, LambdaFunctionURLEvent},
+		// Ultimately check this is a Kong API Gateway event as a last resort.
+		// This is because Kong API Gateway events are a subset of API Gateway events
+		// as of https://github.com/Kong/kong/blob/348c980/kong/plugins/aws-lambda/request-util.lua#L248-L260
+		{isKongAPIGatewayEvent, APIGatewayEvent},
+	}
+)
+
 // GetEventType takes in a payload string and returns an AWSEventType
 // that matches the input payload. Returns `Unknown` if a payload could not be
 // matched to an event.
 func GetEventType(payload map[string]any) AWSEventType {
-	if isAPIGatewayEvent(payload) {
-		return APIGatewayEvent
+	// allow last event type to jump ahead in line since it is likely the event
+	// type is the same between invocations
+	if lastEventChecker != unknownChecker {
+		if lastEventChecker.check(payload) {
+			return lastEventChecker.typ
+		}
 	}
-
-	if isAPIGatewayV2Event(payload) {
-		return APIGatewayV2Event
+	for _, checker := range eventCheckers {
+		if checker == lastEventChecker {
+			continue // typ already checked
+		}
+		if checker.check(payload) {
+			lastEventChecker = checker
+			return checker.typ
+		}
 	}
-
-	if isAPIGatewayWebsocketEvent(payload) {
-		return APIGatewayWebsocketEvent
-	}
-
-	if isAPIGatewayLambdaAuthorizerTokenEvent(payload) {
-		return APIGatewayLambdaAuthorizerTokenEvent
-	}
-
-	if isAPIGatewayLambdaAuthorizerRequestParametersEvent(payload) {
-		return APIGatewayLambdaAuthorizerRequestParametersEvent
-	}
-
-	if isALBEvent(payload) {
-		return ALBEvent
-	}
-
-	if isCloudFrontRequestEvent(payload) {
-		return CloudFrontRequestEvent
-	}
-
-	if isCloudwatchEvent(payload) {
-		return CloudWatchEvent
-	}
-
-	if isCloudwatchLogsEvent(payload) {
-		return CloudWatchLogsEvent
-	}
-
-	if isDynamoDBStreamEvent(payload) {
-		return DynamoDBStreamEvent
-	}
-
-	if isKinesisStreamEvent(payload) {
-		return KinesisStreamEvent
-	}
-
-	if isS3Event(payload) {
-		return S3Event
-	}
-
-	if isSNSEvent(payload) {
-		return SNSEvent
-	}
-
-	if isSNSSQSEvent(payload) {
-		return SNSSQSEvent
-	}
-
-	if isSQSEvent(payload) {
-		return SQSEvent
-	}
-
-	if isAppSyncResolverEvent(payload) {
-		return AppSyncResolverEvent
-	}
-
-	if isEventBridgeEvent(payload) {
-		return EventBridgeEvent
-	}
-
-	if isLambdaFunctionURLEvent(payload) {
-		return LambdaFunctionURLEvent
-	}
-
-	// Ultimately check this is a Kong API Gateway event as a last resort.
-	// This is because Kong API Gateway events are a subset of API Gateway events
-	// as of https://github.com/Kong/kong/blob/348c980/kong/plugins/aws-lambda/request-util.lua#L248-L260
-	if isKongAPIGatewayEvent(payload) {
-		return APIGatewayEvent
-	}
-
+	lastEventChecker = unknownChecker
 	return Unknown
 }