… pkg/serverless/trace/testdata/event_samples/empty.json