<!--
* New contributors are highly encouraged to read our
  [CONTRIBUTING](/CONTRIBUTING.md) documentation.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's change.

-->
### What does this PR do?

Most universal instrumentation languages (Dotnet, Golang, Java, Ruby) are strongly typed, that is all but Ruby. Therefore, there is a very low chance that the event type will change between invocations.

This PR determines the event type for each request, then caches it. Subsequent requests will then first check to see if they match that type. If it matches, return it, otherwise check against the other remaining types.

I have a relatively high level of confidence in this PR as the test I wrote checks every combination of pairs of event types and ensures that when the type is cached from the first, the second still returns the correct type.

<!--
* A brief description of the change being made with this pull request.
* If the description here cannot be expressed in a succinct form, consider
  opening multiple pull requests instead of a single one.
-->

### Motivation

After adding the `BenchmarkStartEndInvocation` benchmark, it's easier for us to see what impact a PR has on overhead.

This pull request reduces start/end-invocation time by roughly 4%, with greater impact occurring on SNS/SQS events.

Example comparison before/after:

```
goos: darwin
goarch: arm64
pkg: github.com/DataDog/datadog-agent/pkg/serverless/daemon
                                      │      old      │                 new                  │
                                      │    sec/op     │    sec/op     vs base                │
api-gateway-appsec.json                  102.3µ ±  3%   105.7µ ±  6%        ~ (p=0.052 n=10)
api-gateway-kong-appsec.json             61.85µ ±  1%   60.72µ ±  1%   -1.84% (p=0.001 n=10)
api-gateway-kong.json                    58.59µ ±  6%   57.55µ ±  4%        ~ (p=0.089 n=10)
api-gateway-non-proxy-async.json        101.80µ ±  2%   98.75µ ±  1%   -3.00% (p=0.000 n=10)
api-gateway-non-proxy.json              101.22µ ±  1%   98.03µ ±  2%   -3.15% (p=0.005 n=10)
api-gateway-websocket-connect.json       66.15µ ±  2%   63.37µ ±  4%   -4.21% (p=0.015 n=10)
api-gateway-websocket-default.json       54.72µ ±  3%   54.41µ ±  3%        ~ (p=0.393 n=10)
api-gateway-websocket-disconnect.json    54.38µ ±  1%   54.57µ ± 50%        ~ (p=0.280 n=10)
api-gateway.json                         113.1µ ±  1%   115.4µ ±  6%   +2.07% (p=0.023 n=10)
application-load-balancer.json           45.30µ ±  1%   45.91µ ±  0%   +1.36% (p=0.003 n=10)
cloudfront.json                          37.88µ ±  7%   37.72µ ±  4%        ~ (p=0.912 n=10)
cloudwatch-events.json                   27.07µ ±  2%   27.84µ ±  2%   +2.84% (p=0.000 n=10)
cloudwatch-logs.json                     37.80µ ± 12%   37.23µ ± 12%        ~ (p=1.000 n=10)
custom.json                              20.00µ ± 27%   18.26µ ±  8%   -8.72% (p=0.043 n=10)
dynamodb.json                            113.0µ ±  2%   116.1µ ±  1%   +2.70% (p=0.000 n=10)
eventbridge-custom.json                  31.12µ ±  2%   31.54µ ±  2%   +1.35% (p=0.029 n=10)
http-api.json                            56.20µ ±  1%   60.06µ ±  2%   +6.87% (p=0.000 n=10)
kinesis-batch.json                       71.80µ ±  4%   76.87µ ± 13%   +7.06% (p=0.002 n=10)
kinesis.json                             51.62µ ±  1%   51.99µ ±  1%        ~ (p=0.105 n=10)
s3.json                                  53.82µ ±  5%   55.28µ ±  2%   +2.72% (p=0.019 n=10)
sns-batch.json                           101.0µ ±  1%   101.0µ ±  2%        ~ (p=0.912 n=10)
sns.json                                 62.53µ ±  2%   63.53µ ±  1%        ~ (p=0.075 n=10)
snssqs.json                             102.57µ ±  7%   94.86µ ±  3%   -7.51% (p=0.002 n=10)
snssqs_no_dd_context.json               101.69µ ±  5%   96.35µ ±  2%   -5.25% (p=0.000 n=10)
sqs-aws-header.json                      52.88µ ± 20%   46.71µ ±  2%  -11.68% (p=0.000 n=10)
sqs-batch.json                          115.60µ ±  5%   81.02µ ±  1%  -29.91% (p=0.000 n=10)
sqs.json                                 81.09µ ±  8%   55.33µ ±  1%  -31.78% (p=0.000 n=10)
sqs_no_dd_context.json                   76.60µ ±  7%   55.24µ ±  1%  -27.89% (p=0.000 n=10)
geomean                                  63.46µ         60.67µ         -4.39%

                                      │      old      │                 new                  │
                                      │     B/op      │     B/op       vs base               │
api-gateway-appsec.json                 53.83Ki ±  2%   53.73Ki ±  2%       ~ (p=0.529 n=10)
api-gateway-kong-appsec.json            30.46Ki ± 11%   30.40Ki ±  9%       ~ (p=0.138 n=10)
api-gateway-kong.json                   27.78Ki ± 22%   27.72Ki ± 22%  -0.21% (p=0.013 n=10)
api-gateway-non-proxy-async.json        58.01Ki ±  0%   57.93Ki ±  0%  -0.15% (p=0.000 n=10)
api-gateway-non-proxy.json              56.57Ki ±  0%   56.49Ki ±  0%  -0.14% (p=0.006 n=10)
api-gateway-websocket-connect.json      35.58Ki ±  0%   35.51Ki ±  0%  -0.19% (p=0.014 n=10)
api-gateway-websocket-default.json      29.34Ki ±  0%   29.29Ki ±  0%  -0.17% (p=0.018 n=10)
api-gateway-websocket-disconnect.json   28.87Ki ±  0%   28.86Ki ±  0%       ~ (p=0.159 n=10)
api-gateway.json                        62.45Ki ±  0%   62.41Ki ±  0%  -0.07% (p=0.002 n=10)
application-load-balancer.json          24.75Ki ±  0%   23.82Ki ±  0%  -3.79% (p=0.000 n=10)
cloudfront.json                         18.57Ki ±  0%   18.55Ki ±  0%  -0.09% (p=0.001 n=10)
cloudwatch-events.json                  12.58Ki ±  0%   12.55Ki ±  0%  -0.26% (p=0.001 n=10)
cloudwatch-logs.json                    54.55Ki ±  0%   54.54Ki ±  0%       ~ (p=0.269 n=10)
custom.json                             9.690Ki ±  0%   9.671Ki ±  0%  -0.20% (p=0.001 n=10)
dynamodb.json                           59.51Ki ±  0%   59.48Ki ±  0%  -0.05% (p=0.000 n=10)
eventbridge-custom.json                 13.76Ki ±  0%   13.72Ki ±  0%  -0.23% (p=0.000 n=10)
http-api.json                           28.44Ki ±  0%   28.46Ki ±  0%       ~ (p=0.184 n=10)
kinesis-batch.json                      35.92Ki ±  0%   35.76Ki ±  0%  -0.43% (p=0.001 n=10)
kinesis.json                            25.92Ki ±  0%   25.76Ki ±  0%  -0.61% (p=0.000 n=10)
s3.json                                 21.86Ki ±  0%   21.82Ki ±  0%  -0.17% (p=0.000 n=10)
sns-batch.json                          49.03Ki ±  0%   48.93Ki ±  0%  -0.20% (p=0.000 n=10)
sns.json                                32.30Ki ±  0%   32.21Ki ±  0%  -0.29% (p=0.000 n=10)
snssqs.json                             53.58Ki ±  0%   53.38Ki ±  0%       ~ (p=0.075 n=10)
snssqs_no_dd_context.json               53.40Ki ±  0%   52.26Ki ±  0%  -2.13% (p=0.000 n=10)
sqs-aws-header.json                     24.15Ki ±  1%   24.23Ki ±  1%  +0.30% (p=0.035 n=10)
sqs-batch.json                          47.87Ki ±  0%   48.11Ki ±  0%  +0.51% (p=0.000 n=10)
sqs.json                                31.08Ki ±  0%   31.15Ki ±  0%       ~ (p=0.143 n=10)
sqs_no_dd_context.json                  28.74Ki ±  0%   28.59Ki ±  0%  -0.54% (p=0.000 n=10)
geomean                                 32.36Ki         32.25Ki        -0.35%

                                      │    old     │                 new                 │
                                      │ allocs/op  │ allocs/op   vs base                 │
api-gateway-appsec.json                 680.0 ± 0%   680.0 ± 0%       ~ (p=1.000 n=10) ¹
api-gateway-kong-appsec.json            487.5 ± 0%   487.0 ± 0%  -0.10% (p=0.033 n=10)
api-gateway-kong.json                   465.0 ± 0%   465.0 ± 0%       ~ (p=0.474 n=10)
api-gateway-non-proxy-async.json        698.0 ± 0%   697.5 ± 0%  -0.07% (p=0.033 n=10)
api-gateway-non-proxy.json              689.0 ± 0%   688.0 ± 0%       ~ (p=0.070 n=10)
api-gateway-websocket-connect.json      465.0 ± 0%   465.0 ± 0%       ~ (p=0.303 n=10)
api-gateway-websocket-default.json      376.0 ± 0%   376.0 ± 0%       ~ (p=1.000 n=10)
api-gateway-websocket-disconnect.json   367.0 ± 0%   367.0 ± 0%       ~ (p=1.000 n=10)
api-gateway.json                        791.0 ± 0%   791.0 ± 0%       ~ (p=1.000 n=10) ¹
application-load-balancer.json          337.0 ± 0%   336.0 ± 0%  -0.30% (p=0.000 n=10)
cloudfront.json                         246.0 ± 0%   246.0 ± 0%       ~ (p=1.000 n=10) ¹
cloudwatch-events.json                  227.0 ± 0%   227.0 ± 0%       ~ (p=1.000 n=10) ¹
cloudwatch-logs.json                    225.0 ± 0%   225.0 ± 0%       ~ (p=1.000 n=10) ¹
custom.json                             172.0 ± 0%   172.0 ± 0%       ~ (p=1.000 n=10) ¹
dynamodb.json                           698.0 ± 0%   698.0 ± 0%       ~ (p=1.000 n=10)
eventbridge-custom.json                 242.0 ± 0%   242.0 ± 0%       ~ (p=1.000 n=10) ¹
http-api.json                           405.0 ± 0%   405.0 ± 0%       ~ (p=1.000 n=10)
kinesis-batch.json                      411.0 ± 0%   411.0 ± 0%       ~ (p=1.000 n=10) ¹
kinesis.json                            352.0 ± 0%   352.0 ± 0%       ~ (p=1.000 n=10) ¹
s3.json                                 306.0 ± 0%   306.0 ± 0%       ~ (p=1.000 n=10) ¹
sns-batch.json                          442.0 ± 0%   442.0 ± 0%       ~ (p=1.000 n=10) ¹
sns.json                                369.0 ± 0%   369.0 ± 0%       ~ (p=0.582 n=10)
snssqs.json                             392.0 ± 0%   392.0 ± 0%       ~ (p=0.087 n=10)
snssqs_no_dd_context.json               459.0 ± 0%   458.0 ± 0%  -0.22% (p=0.000 n=10)
sqs-aws-header.json                     271.0 ± 0%   271.0 ± 0%       ~ (p=1.000 n=10) ¹
sqs-batch.json                          432.0 ± 0%   432.0 ± 0%       ~ (p=0.474 n=10)
sqs.json                                332.0 ± 0%   332.0 ± 0%       ~ (p=1.000 n=10) ¹
sqs_no_dd_context.json                  361.0 ± 0%   361.0 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                 389.5        389.4       -0.03%
¹ all samples are equal
```

<!--
* What inspired you to submit this pull request?
* Link any related GitHub issues or PRs here.
-->

### Additional Notes

<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the proposed
  version was chosen.
-->

### Possible Drawbacks / Trade-offs

<!--
* What are the possible side-effects or negative impacts of the code change?
-->

### Describe how to test/QA your changes

<!--
* Write here in detail or link to detailed instructions on how this change can
  be tested/QAd/validated, including any environment setup.
-->

### Reviewer's Checklist
<!--
* Authors can use this list as a reference to ensure that there are no problems
  during the review but the signing off is to be done by the reviewer(s).

Note: Adding GitHub labels is only possible for contributors with write access.
-->

- [ ] If known, an appropriate milestone has been selected; otherwise the `Triage` milestone is set.
- [ ] Use the `major_change` label if your change either has a major impact on the code base, is impacting multiple teams or is changing important well-established internals of the Agent. This label will be use during QA to make sure each team pay extra attention to the changed behavior. For any customer facing change use a releasenote.
- [ ] A [release note](https://github.com/DataDog/datadog-agent/blob/main/docs/dev/contributing.md#reno) has been added or the `changelog/no-changelog` label has been applied.
- [ ] Changed code has automated tests for its functionality.
- [ ] Adequate QA/testing plan information is provided if the `qa/skip-qa` label is not applied.
- [ ] At least one `team/..` label has been applied, indicating the team(s) that should QA this change.
- [ ] If applicable, docs team has been notified or [an issue has been opened on the documentation repo](https://github.com/DataDog/documentation/issues/new).
- [ ] If applicable, the `need-change/operator` and `need-change/helm` labels have been applied.
- [ ] If applicable, the `k8s/<min-version>` label, indicating the lowest Kubernetes version compatible with this feature.
- [ ] If applicable, the [config template](https://github.com/DataDog/datadog-agent/blob/main/pkg/config/config_template.yaml) has been updated.